DataTable models in C#
I've written this set of C# models to help developers out when using the
JQuery DataTables plugin (version 1.10 as of now)

It's (I think) a fairly straight forward implementation.
The namespace: DataTables.Models.Input holds all the necessary classes
that are sent to the server by the plugin.
The namespace: DataTables.Models.Output holds the classes that are sent
back to the client.

I'll soon post an example using the Output namespace. :D